Refactor common methods in event-aggregator services
