+++ packages/core/src/event-aggregator/gql-service.ts
@@ -5,8 +5,7 @@ import {
   type ObservableSubscription,
   gql,
 } from "@apollo/client";
-import Emittery from "emittery";
-import { type Hex, Address, decodeEventLog } from "viem";
+import { type Hex, Address } from "viem";
 
 import type { LogFilterName } from "@/build/handlers";
 import type { LogEventMetadata, LogFilter } from "@/config/logFilters";
@@ -15,14 +14,8 @@ import type { Common } from "@/Ponder";
 import type { Block } from "@/types/block";
 import type { Log } from "@/types/log";
 import type { Transaction } from "@/types/transaction";
-import { formatShortDate } from "@/utils/date";
 
-import type {
-  EventAggregatorEvents,
-  EventAggregatorMetrics,
-  EventAggregatorService,
-  LogEvent,
-} from "./service";
+import { EventAggregatorService } from "./service";
 
 type Cursor = {
   timestamp: number;
@@ -31,36 +24,9 @@ type Cursor = {
   logIndex: number;
 };
 
-export class GqlEventAggregatorService
-  extends Emittery<EventAggregatorEvents>
-  implements EventAggregatorService
-{
-  private common: Common;
+export class GqlEventAggregatorService extends EventAggregatorService {
   private gqlClient: ApolloClient<NormalizedCacheObject>;
-  private logFilters: LogFilter[];
-  private networks: Network[];
   private subscriptions: ObservableSubscription[] = [];
-
-  // Minimum timestamp at which events are available (across all networks).
-  checkpoint: number;
-  // Minimum finalized timestamp (across all networks).
-  finalityCheckpoint: number;
-
-  // Timestamp at which the historical sync was completed (across all networks).
-  historicalSyncCompletedAt?: number;
-
-  // Per-network event timestamp checkpoints.
-  private networkCheckpoints: Record<
-    number,
-    {
-      isHistoricalSyncComplete: boolean;
-      historicalCheckpoint: number;
-      realtimeCheckpoint: number;
-      finalityCheckpoint: number;
-    }
-  >;
-
-  metrics: EventAggregatorMetrics;
 
   constructor({
     common,
@@ -73,27 +39,15 @@ export class GqlEventAggregatorService
     networks: Network[];
     logFilters: LogFilter[];
   }) {
-    super();
+    super({
+      common,
+      networks,
+      logFilters,
+    });
 
-    this.common = common;
-    this.logFilters = logFilters;
-    this.networks = networks;
     this.metrics = {};
 
     this.gqlClient = gqlClient;
-
-    this.checkpoint = 0;
-    this.finalityCheckpoint = 0;
-
-    this.networkCheckpoints = {};
-    this.networks.forEach((network) => {
-      this.networkCheckpoints[network.chainId] = {
-        isHistoricalSyncComplete: false,
-        historicalCheckpoint: 0,
-        realtimeCheckpoint: 0,
-        finalityCheckpoint: 0,
-      };
-    });
   }
 
   /** Fetches events for all registered log filters between the specified timestamps.
@@ -141,49 +95,7 @@ export class GqlEventAggregatorService
       // Set cursor to fetch next batch of events from indexingClient GQL query
       cursor = metadata.cursor;
 
-      const decodedEvents = events.reduce<LogEvent[]>((acc, event) => {
-        const selector = event.log.topics[0];
-        if (!selector) {
-          throw new Error(
-            `Received an event log with no selector: ${event.log}`
-          );
-        }
-
-        const logEventMetadata =
-          includeLogFilterEvents[event.logFilterName]?.bySelector[selector];
-        if (!logEventMetadata) {
-          throw new Error(
-            `Metadata for event ${event.logFilterName}:${selector} not found in includeLogFilterEvents`
-          );
-        }
-        const { abiItem, safeName } = logEventMetadata;
-
-        try {
-          const decodedLog = decodeEventLog({
-            abi: [abiItem],
-            data: event.log.data,
-            topics: event.log.topics,
-          });
-
-          acc.push({
-            logFilterName: event.logFilterName,
-            eventName: safeName,
-            params: decodedLog.args || {},
-            log: event.log,
-            block: event.block,
-            transaction: event.transaction,
-          });
-        } catch (err) {
-          // TODO: emit a warning here that a log was not decoded.
-          this.common.logger.error({
-            service: "app",
-            msg: `Unable to decode log (skipping it): ${event.log}`,
-            error: err as Error,
-          });
-        }
-
-        return acc;
-      }, []);
+      const decodedEvents = this.decodeEvents(events, includeLogFilterEvents);
 
       yield { events: decodedEvents, metadata };
 
@@ -191,7 +103,7 @@ export class GqlEventAggregatorService
     }
   }
 
-  subscribeToSyncEvents() {
+  override subscribeToSyncEvents() {
     this.subscriptions = [
       this.subscribeGql(
         gql`
@@ -259,130 +171,10 @@ export class GqlEventAggregatorService
     ];
   }
 
-  kill() {
+  override kill() {
     this.subscriptions.forEach((subscription) => subscription.unsubscribe());
     this.clearListeners();
   }
-
-  // TODO: Refactor common methods in event aggregator services
-  handleNewHistoricalCheckpoint = ({
-    chainId,
-    timestamp,
-  }: {
-    chainId: number;
-    timestamp: number;
-  }) => {
-    this.networkCheckpoints[chainId].historicalCheckpoint = timestamp;
-
-    this.common.logger.trace({
-      service: "aggregator",
-      msg: `New historical checkpoint at ${timestamp} [${formatShortDate(
-        timestamp
-      )}] (chainId=${chainId})`,
-    });
-
-    this.recalculateCheckpoint();
-  };
-
-  handleHistoricalSyncComplete = ({ chainId }: { chainId: number }) => {
-    this.networkCheckpoints[chainId].isHistoricalSyncComplete = true;
-    this.recalculateCheckpoint();
-
-    // If every network has completed the historical sync, set the metric.
-    const networkCheckpoints = Object.values(this.networkCheckpoints);
-    if (networkCheckpoints.every((n) => n.isHistoricalSyncComplete)) {
-      const maxHistoricalCheckpoint = Math.max(
-        ...networkCheckpoints.map((n) => n.historicalCheckpoint)
-      );
-      this.historicalSyncCompletedAt = maxHistoricalCheckpoint;
-
-      this.common.logger.debug({
-        service: "aggregator",
-        msg: `Completed historical sync across all networks`,
-      });
-    }
-  };
-
-  handleNewRealtimeCheckpoint = ({
-    chainId,
-    timestamp,
-  }: {
-    chainId: number;
-    timestamp: number;
-  }) => {
-    this.networkCheckpoints[chainId].realtimeCheckpoint = timestamp;
-
-    this.common.logger.trace({
-      service: "aggregator",
-      msg: `New realtime checkpoint at ${timestamp} [${formatShortDate(
-        timestamp
-      )}] (chainId=${chainId})`,
-    });
-
-    this.recalculateCheckpoint();
-  };
-
-  handleNewFinalityCheckpoint = ({
-    chainId,
-    timestamp,
-  }: {
-    chainId: number;
-    timestamp: number;
-  }) => {
-    this.networkCheckpoints[chainId].finalityCheckpoint = timestamp;
-    this.recalculateFinalityCheckpoint();
-  };
-
-  handleReorg = ({
-    commonAncestorTimestamp,
-  }: {
-    commonAncestorTimestamp: number;
-  }) => {
-    this.emit("reorg", { commonAncestorTimestamp });
-  };
-
-  private recalculateCheckpoint = () => {
-    const checkpoints = Object.values(this.networkCheckpoints).map((n) =>
-      n.isHistoricalSyncComplete
-        ? Math.max(n.historicalCheckpoint, n.realtimeCheckpoint)
-        : n.historicalCheckpoint
-    );
-    const newCheckpoint = Math.min(...checkpoints);
-
-    if (newCheckpoint > this.checkpoint) {
-      this.checkpoint = newCheckpoint;
-
-      this.common.logger.trace({
-        service: "aggregator",
-        msg: `New event checkpoint at ${this.checkpoint} [${formatShortDate(
-          this.checkpoint
-        )}]`,
-      });
-
-      this.emit("newCheckpoint", { timestamp: this.checkpoint });
-    }
-  };
-
-  private recalculateFinalityCheckpoint = () => {
-    const newFinalityCheckpoint = Math.min(
-      ...Object.values(this.networkCheckpoints).map((n) => n.finalityCheckpoint)
-    );
-
-    if (newFinalityCheckpoint > this.finalityCheckpoint) {
-      this.finalityCheckpoint = newFinalityCheckpoint;
-
-      this.common.logger.trace({
-        service: "aggregator",
-        msg: `New finality checkpoint at ${
-          this.finalityCheckpoint
-        } [${formatShortDate(this.finalityCheckpoint)}]`,
-      });
-
-      this.emit("newFinalityCheckpoint", {
-        timestamp: this.finalityCheckpoint,
-      });
-    }
-  };
 
   private getLogEvents = async (variables: {
     fromTimestamp: number;
